+++ drink/test-macro/src/lib.rs
@@ -56,7 +56,7 @@ type SynResult<T> = Result<T, syn::Error>;
 /// #[drink::test]
 /// fn testcase(mut session: Session<MinimalSandbox>) {
 ///     session
-///         .deploy_bundle(&get_bundle(), "new", NO_ARGS, NO_SALT, NO_ENDOWMENT)
+///         .deploy_bundle(&get_bundle(), "new", NO_ARGS, NO_SALT, None)
 ///         .unwrap();
 /// }
 /// ```
@@ -78,6 +78,7 @@ fn test_internal(attr: TokenStream2, item: TokenStream2) -> SynResult<TokenStrea
     let item_fn = syn::parse2::<ItemFn>(item)?;
     let macro_args = TestAttributes::from_list(&NestedMeta::parse_meta_list(attr)?)?;
 
+    // TODO: why do we build the contracts again?
     build_contracts();
 
     let fn_vis = item_fn.vis;
@@ -137,8 +138,8 @@ fn test_internal(attr: TokenStream2, item: TokenStream2) -> SynResult<TokenStrea
 ///
 /// fn testcase() {
 ///     Session::<MinimalSandbox>::default()
-///         .deploy_bundle_and(BundleProvider::local()?, "new", NO_ARGS, NO_SALT, NO_ENDOWMENT)
-///         .deploy_bundle_and(BundleProvider::AnotherContract.bundle()?, "new", NO_ARGS, NO_SALT, NO_ENDOWMENT)
+///         .deploy_bundle_and(BundleProvider::local()?, "new", NO_ARGS, NO_SALT, None)
+///         .deploy_bundle_and(BundleProvider::AnotherContract.bundle()?, "new", NO_ARGS, NO_SALT, None)
 ///         .unwrap();
 /// }
 /// ```