+++ examples/contract-events/lib.rs
@@ -37,7 +37,7 @@ mod flipper {
 mod tests {
     use std::error::Error;
 
-    use drink::session::{Session, NO_ARGS, NO_ENDOWMENT};
+    use drink::session::{Session, NO_ARGS};
 
     #[drink::contract_bundle_provider]
     enum BundleProvider {}
@@ -47,8 +47,8 @@ mod tests {
         let bundle = BundleProvider::local()?;
 
         // Firstly, we deploy the contract and call its `flip` method.
-        session.deploy_bundle(bundle.clone(), "new", &["false"], vec![], NO_ENDOWMENT)?;
-        session.call("flip", NO_ARGS, NO_ENDOWMENT)??;
+        session.deploy_bundle(bundle.clone(), "new", &["false"], vec![], None)?;
+        session.call("flip", NO_ARGS, None)??;
 
         // Now we can inspect the emitted events.
         let record = session.record();