+++ examples/mocking/lib.rs
@@ -41,7 +41,7 @@ mod tests {
 
     use drink::{
         mock_message,
-        session::{mocking_api::MockingApi, Session, NO_ARGS, NO_ENDOWMENT, NO_SALT},
+        session::{mocking_api::MockingApi, Session, NO_ARGS, NO_SALT},
         ContractMock,
     };
 
@@ -63,7 +63,7 @@ mod tests {
         // Now, we can deploy our proper contract and verify its behavior.
         let result: (u8, u8) = session
             .deploy_bundle_and(BundleProvider::local()?, "new", NO_ARGS, NO_SALT, None)?
-            .call_and("forward_call", &[mock_address.to_string()], NO_ENDOWMENT)?
+            .call_and("forward_call", &[mock_address.to_string()], None)?
             .record()
             .last_call_return_decoded()?
             .expect("Call was successful");